backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
apscheduler